Rename updateName component to better match new features
