+++ public/locales/en/common.json
@@ -25,7 +25,7 @@
       "remove_avatar": "Remove avatar",
       "select_avatar": "Select picture",
       "zoom": "Zoom",
-      "title": "Edit name"
+      "title": "Edit details"
     },
     "follow_on_x": "Follow us on X",
     "import_from_splitwise": "Import from Splitwise",
